支持回调函数。

+++ src/ECommon/Remoting/RemotingMessage.cs
@@ -3,10 +3,10 @@
     public class RemotingMessage
     {
         public long Sequence { get; private set; }
-        public int Code { get; private set; }
+        public short Code { get; private set; }
         public byte[] Body { get; private set; }
 
-        public RemotingMessage(int code, long sequence, byte[] body)
+        public RemotingMessage(short code, long sequence, byte[] body)
         {
             Code = code;
             Sequence = sequence;
